Can now switch between states.
On first attempt the memory leak was disgusting. However the offenders were found:
1) Unsubscribe events e.g. MainGame.Exiting.
2) AntiShadow and SelfShadows are now cached in static Dictionaries (GraphicsHelper class) the first time they are made.

Create overlay now happens after deserialization, such that button states match. There is still an issue when switching states whereby the GhostMarker does not update.

Currently there is a chance that Draw and DrawOverlay of the old state are called when switching states.
Wrapping the block in a conditional does nothing which is very odd.
At worst, we will need to null check objects in Draw and DrawOverlay (in general this is already dealt with via a foreach loop which simply has a count of zero).

Signed-off-by: Andrew <54269534+awu2147@users.noreply.github.com>